+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # @azure-tools/rest-api-diff
 
+## 0.2.2 (2025-04-11)
+
+- Removed `--group-violations` options. Grouping is now always performed.
+- Fixed issue where `allOf` was not removed from the output.
+- Fixes issue where infinite loop could occur with inheritance chains.
+
 ## 0.2.1 (2025-04-09)
 
 - Added `--suppressions` option to point to a file containing point suppressions of violations.